+++ rest-api-finances/src/main/java/com/abnerhs/rest_api_finances/dto/TransactionRequestDTO.java
@@ -34,5 +34,6 @@ public record TransactionRequestDTO(
         Boolean isClearedByInvoice,
         LocalDate dueDate,
         LocalDate paymentDate,
-        PaymentStatus paymentStatus
+        PaymentStatus paymentStatus,
+        BillingDocumentRequestDTO billingDocument
 ) {}