+++ rest-api-finances/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.4.0](https://github.com/AbnerHS/FinancesFullstack/compare/v0.3.0...v0.4.0) (2026-04-01)
+
+
+### Features
+
+* upload billing document in transaction ([#25](https://github.com/AbnerHS/FinancesFullstack/issues/25)) ([6ea5761](https://github.com/AbnerHS/FinancesFullstack/commit/6ea57610082a864bd883e39f823a32382e36aef1))
+
+## [0.3.0](https://github.com/AbnerHS/FinancesFullstack/compare/v0.2.1...v0.3.0) (2026-03-25)
+
+
+### Features
+
+* improve transaction workflow and payment tracking UI ([#22](https://github.com/AbnerHS/FinancesFullstack/issues/22)) ([6a0a70a](https://github.com/AbnerHS/FinancesFullstack/commit/6a0a70a49444fb9c82bc1c2aa1f3813942d01ec4))
+
 ## [0.2.1](https://github.com/AbnerHS/FinancesFullstack/compare/v0.2.0...v0.2.1) (2026-03-25)